fix(ft): use native pointer-width fighter data

+++ src/melee/ft/ftdata.c
@@ -1656,7 +1656,7 @@ void ftData_80085A14(FighterKind kind)
     void* sp18;
     void* a_head;
     ftData* temp_r27 = gFtDataList[kind];
-    u32 temp_r0;
+    uintptr_t temp_r0;
     int i;
     u8 _[4];
     size_t sp10;
@@ -1695,12 +1695,12 @@ void ftData_80085B10(Fighter* fp)
 
 void ftData_80085B98(Fighter* fp, int arg1, int arg2)
 {
-    u32 temp_r30;
+    uintptr_t temp_r30;
     int i;
-    u32 temp_r0;
+    uintptr_t temp_r0;
     struct Fighter_WaitAnimData* temp_r3;
 
-    temp_r30 = (u32) ftData_UnkIntPairs[fp->kind].data;
+    temp_r30 = (uintptr_t) ftData_UnkIntPairs[fp->kind].data;
     fp->x59C = HSD_ObjAlloc(&fighter_x59C_alloc_data);
     fp->x5A0 = HSD_ObjAlloc(&fighter_x59C_alloc_data);
     fp->x5A4 = 0;
@@ -1734,17 +1734,17 @@ void ftData_80085CD8(Fighter* fp, Fighter* arg1, int msid)
     s32 temp_ret_2;
     struct Fighter_x59C_t* temp_r4;
     struct Fighter_WaitAnimData* temp_r3;
-    u32 temp_r3_2;
-    u32 temp_r4_2;
+    uintptr_t temp_r3_2;
+    uintptr_t temp_r4_2;
 
     if (msid < arg1->x58C) {
         temp_r3 = (struct Fighter_WaitAnimData*) ftData_80085FD4(arg1, msid);
         temp_r3_2 = temp_r3->x14;
-        if (temp_r3_2 != (u32) fp->x5A4) {
+        if (temp_r3_2 != (uintptr_t) fp->x5A4) {
             if (temp_r3_2 != 0) {
                 temp_r3_3 = ftData_80086060(fp);
                 if ((temp_r3_3 != NULL) &&
-                    (temp_r3->x14 == (u32) temp_r3_3->x5A4))
+                    (temp_r3->x14 == (uintptr_t) temp_r3_3->x5A4))
                 {
                     memcpy(fp->x59C, temp_r3_3->x59C, temp_r3->x8);
                     temp_r4 = fp->x59C;
@@ -1761,7 +1761,7 @@ void ftData_80085CD8(Fighter* fp, Fighter* arg1, int msid)
                         lbArq_80014BD0(temp_r4_2, fp->x59C,
                                        OSRoundUp32B(temp_r3->x8), 0, 0);
                     } else {
-                        memcpy(fp->x59C, (void*) temp_r4_2, temp_r3->x8);
+                        memcpy(fp->x59C, (void*) (uintptr_t) temp_r4_2, temp_r3->x8);
                     }
                     temp_ret_2 =
                         HSD_ArchiveParse(&sp14, fp->x59C->x0, temp_r3->x8);
@@ -1787,17 +1787,17 @@ FigaTree* ftData_80085E50(Fighter* arg0, int msid)
     int temp_ret_2;
     struct Fighter_x59C_t* temp_r4;
     struct ftData_80085FD4_ret* temp_r3;
-    u32 temp_r3_2;
-    u32 temp_r4_2;
+    uintptr_t temp_r3_2;
+    uintptr_t temp_r4_2;
 
     if (msid < arg0->x58C) {
         temp_r3 = ftData_80085FD4(arg0, msid);
         temp_r3_2 = temp_r3->x14;
-        if (temp_r3_2 != (u32) arg0->x5A8) {
+        if (temp_r3_2 != (uintptr_t) arg0->x5A8) {
             if (temp_r3_2 != 0) {
                 temp_r3_3 = ftData_80086060(arg0);
                 if ((temp_r3_3 != NULL) &&
-                    (temp_r3->x14 == (u32) temp_r3_3->x5A4))
+                    (temp_r3->x14 == (uintptr_t) temp_r3_3->x5A4))
                 {
                     memcpy(arg0->x59C, temp_r3_3->x59C, temp_r3->x8);
                     temp_r4 = arg0->x59C;
@@ -1814,7 +1814,7 @@ FigaTree* ftData_80085E50(Fighter* arg0, int msid)
                         lbArq_80014BD0(temp_r4_2, arg0->x5A0,
                                        OSRoundUp32B(temp_r3->x8), 0, 0);
                     } else {
-                        memcpy(arg0->x5A0, (void*) temp_r4_2, temp_r3->x8);
+                        memcpy(arg0->x5A0, (void*) (uintptr_t) temp_r4_2, temp_r3->x8);
                     }
                     temp_ret_2 =
                         HSD_ArchiveParse(&sp10, arg0->x5A0->x0, temp_r3->x8);
